[Refactor ⚙️] TanStack Query 컨벤션에 적합하도록 리팩터링 및 폴더 구조 정리
